SK-1353: remove Redux: verifiedSignature
